Migrate to Next.js
The project was initiated using Django. It was proving to be more
troublesome than it should have been.
Seeing that the focus of the project is to implement the adaptive model,
the decision was made to use Next.js instead - since there is some
familiarity with implementing and hosting it.